changed mkl linking to use -lmkl_rt and use mkl_set_num_threads_local in script. This way the user does not have to set the LD_PRELOAD path and we are guaranteed to not have nested parallel computation

+++ cython/cmat_c_utils.pxd
@@ -4,10 +4,10 @@ cdef extern from "optweight_mat_c_utils.h":
                           const float lim,
                           const float lim0,
                           const int nsamp,
-                          const long long int ncomp);
+                          const int ncomp);
     void _eigpow_core_rdp(double *imat,
                           const double power,
                           const double lim,
                           const double lim0,
                           const int nsamp,
-                          const long long int ncomp);			    
+                          const int ncomp);
