+++ include/optweight_mat_c_utils.h
@@ -13,8 +13,6 @@
  */
 
 void _eigpow_core_rsp(float *imat, const float power, const float lim,
-		      const float lim0, const int nsamp,
-		      const long long int ncomp);
+		      const float lim0, const int nsamp, const int ncomp);
 void _eigpow_core_rdp(double *imat, const double power, const double lim,
-		      const double lim0, const int nsamp,
-		      const long long int ncomp);
+		      const double lim0, const int nsamp, const int ncomp);
